Adding discrepancy finder script via upload
I am attempting to refactor the script. I have been developing the script since November. The new, shorter, and refactored one is a version developed in tandem over Jan 3rd and 5th, 2024 by myself, Rich Perez, and Reion Quamina. This shorter version is being uploaded now. Many thanks to Reion for helping develop the refactored structure and the findProblems() function. Thanks to the Agliant survey team for helping develop the logic of finding discrepancies -- pointing out helpful nuances and practices with the data: notable mentions include: Khadejah Harrison, Morgan Mendez, Tanjim Khan, Jesupemi Akanji, Ibrahima, Quedraogo, Yostinn Minaya, Jeniffer Carela, Bill Zhou, Caesar Olivas, Robert Borelli Jr, Gary Gray, Astrid Ovalles, Alicia Mitchell, and Felix Baldomero Jr. Others have helped to lesser capacities, but I offer my gratitude nonetheless.

Currently, the script tells you the ID of a previous duplicate. This is minimal functionality with respect to the scope of discrepancies to look for and bugs to fix in generating these logics.

This version needs to be fixed in several ways:
It does not look for flaws in the current standard of designation
It does not highlight cells of a spreadsheet.
It does not yet complete the logics of other functions in the original script: flagging reasons, incorrect designator, output to CSV, etc.

There's a warning that load_workbook() on line 26 is supposed to have an argument, but our intention is to create a new one if it doesn't load. 

Once I get the highlight_duplicates function working, I want to break it up into its functions: highlight_duplicate_designators(), highlight_incorrect_designators(), highlight_designators_out_of_sequence(), etc. Each function will have logic for finding the type of data to highlight.
Therefore, highlight_duplicate() will eventually become a part of highlight() and will be called highlight_duplicate_designators()
Highlight is an overall function that will output a spreadsheet with updated, marked, and styled cells. I haven't decided whether or not it's better to update an existing sheet or create a blank sheet and update the blank one. 

At this time, the code is only being developed to highlight duplicate IDs of device records
The code is designed to update an already existing sheet separate from input file; in practice, the one being updated is a copy of the input file in order to keep an untouched, working version of the data.

The structure for the missing logic is largely present in the main script itself.

This code is developed under a GNU License you can refer to in the repository.

GNU License Key Points:
- Anyone who distributes a work based on the software must also make the source code available under the same terms. This ensures that the software remains free and open.
- The GPL is designed to uphold the principles of free software, as defined by the FSF. This includes the freedom to run, study, modify, and distribute the software.
- If you distribute a program that is covered by the GPL, you must also provide the complete corresponding source code. This ensures that users have access to the source code, allowing them to understand and modify the software.
- Any modifications or derivative works based on GPL-licensed software must also be licensed under the GPL. This extends the copyleft requirements to ensure that modifications remain open and freely available.
- The GPL prohibits adding any further restrictions to the terms it imposes. This means that you cannot impose additional restrictions on recipients beyond what the GPL allows.
- The GPL is designed to be compatible with other GPL-licensed software. This allows developers to create projects that combine multiple GPL-licensed components.
- The GPL has different versions, such as GPL-2.0 and GPL-3.0. Each version includes specific terms, and developers can choose the version they prefer for their projects.