Uploading another broken file
This script is close but falls shy.

Reion also removed functionality and changed the format to dictionaries. I now have to change all related functionality to match dictionaries to restore old features and implement future ones accordingly.

Not the best move to remove and reformat all that functionality.

He removed my sequence check logic which used regex to get valid floor names. That was a beast to figure out.